requirements.txt
atomicwrites==1.2.1
attrs==18.2.0
certifi==2019.3.9
chardet==3.0.4
Click==7.0
coverage==4.5.1
coveralls==1.5.0
docopt==0.6.2
flasgger==0.9.1
Flask==1.0.2
Flask-Cors==3.0.6
Flask-Gunicorn==0.1.1
Flask-JWT-Extended==3.13.0
gunicorn==19.9.0
idna==2.7
itsdangerous==0.24
Jinja2==2.10.1
jsonschema==2.6.0
MarkupSafe==1.0
mistune==0.8.3
more-itertools==4.3.0
nose==1.3.7
pluggy==0.7.1
psycopg2==2.7.5
psycopg2-binary==2.7.5
py==1.6.0
PyJWT==1.6.4
pytest==3.8.1
pytest-cov==2.6.0
python-dotenv==0.10.3
PyYAML==5.1.1
requests==2.22.0
six==1.11.0
urllib3==1.25.3
Werkzeug==0.14.1
